moving KineticsDatabase to rmgpy.data.kinetics.base from __init__
The goal is to remove stuff from __init__ as much as possible.

- moved KineticsDatabase to a new module 'database' in rmgpy.data.kinetics
- updated imports in RMG-Py.
- replaced import * statements by specific classes.
- moved all import statements to the top.
- replaced a relative import (from kinetics) by an absolute one.
- correct import in kineticsTest was failing because KineticsDatabase was moved.
