+++ rmgpy/chemkin.py
@@ -44,8 +44,10 @@ from rmg.model import Species
 from thermo import NASAPolynomial, NASA
 import rmgpy.constants as constants
 from quantity import Quantity
-from data.base import Entry
-from data.kinetics import TemplateReaction, LibraryReaction
+from data.base import Entry 
+from rmgpy.data.kinetics.library import LibraryReaction, KineticsLibrary
+from rmgpy.data.kinetics.family import TemplateReaction, KineticsFamily
+from rmgpy.rmg.pdep import PDepReaction, PDepNetwork
 from rmg.pdep import PDepReaction
 from rmgpy.molecule import Molecule
 from rmgpy.transport import TransportData
@@ -527,8 +529,6 @@ def readReactionComments(reaction, comments, read = True):
     comments. If `read` if False, the reaction is returned as an "Unclassified"
     LibraryReaction.
     """
-    from rmgpy.data.kinetics import LibraryReaction,TemplateReaction, KineticsLibrary, KineticsFamily
-    from rmgpy.rmg.pdep import PDepReaction, PDepNetwork
     
     if read == False:
         # The chemkin file was not generated by either RMG-Py or RMG-Java, thus, there should be no parsing
